Completed Soul path finding to avoid obstacles such as walls and borders. Code is present in the AI folder.
AI Data: Stores this entity's information to be passed to other components
Detectors: Detect if objects in specified layer is present. If true, stores objects in AI Data
SteeringBehaviours: Using the collision objects present in AI Data, vector distances are calculated to determine dangers and interests within AI's given radius.
The steering behaviour can then be used by AI's movement system to guide it out of places like corners when necessary

Also added IsSoul and SoulFire to check if an object is a soul. if it is, soul fire objects on the map instantly destroy the soul when put in contact